create token model

+++ DocShareAPI/Controllers/CollectionsController.cs
@@ -19,7 +19,7 @@ namespace DocShareAPI.Controllers
         }
 
         // 1. Thêm mới một bộ sưu tập
-        [HttpPost]
+        [HttpPost("create-collection")]
         public async Task<IActionResult> CreateCollection([FromBody] CollectionDTO collection)
         {
             //Kiểm tra token
@@ -49,9 +49,15 @@ namespace DocShareAPI.Controllers
         }
 
         // 2. Cập nhật một bộ sưu tập
-        [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateCollection(int id, [FromBody] Collections updatedCollection)
+        [HttpPut("update/{id}")]
+        public async Task<IActionResult> UpdateCollection(int id, [FromBody] CollectionDTO updatedCollection)
         {
+            //Kiểm tra token
+            var decodedToken = HttpContext.Items["DecodedToken"] as DecodedTokenResponse;
+            if (decodedToken == null)
+            {
+                return Unauthorized();
+            }
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -68,6 +74,10 @@ namespace DocShareAPI.Controllers
             collection.Description = updatedCollection.Description;
             collection.is_public = updatedCollection.is_public;
 
+            if (collection.user_id != decodedToken.userID && decodedToken.roleID != "admin")
+            {
+                return StatusCode(401, "Bạn không thể cập nhật bộ sưu tập này");
+            }
             _context.COLLECTIONS.Update(collection);
             await _context.SaveChangesAsync();
 
@@ -75,23 +85,32 @@ namespace DocShareAPI.Controllers
         }
 
         // 3. Xóa một bộ sưu tập
-        [HttpDelete("{id}")]
+        [HttpDelete("delete")]
         public async Task<IActionResult> DeleteCollection(int id)
         {
+            //Kiểm tra token
+            var decodedToken = HttpContext.Items["DecodedToken"] as DecodedTokenResponse;
+            if (decodedToken == null)
+            {
+                return Unauthorized();
+            }
             var collection = await _context.COLLECTIONS.FindAsync(id);
             if (collection == null)
             {
                 return NotFound(new { message = $"Collection with ID {id} not found." });
             }
-
+            if (collection.user_id != decodedToken.userID && decodedToken.roleID != "admin")
+            {
+                return StatusCode(401,"Bạn không thể xóa bộ sưu tập này");
+            }
             _context.COLLECTIONS.Remove(collection);
             await _context.SaveChangesAsync();
 
-            return NoContent(); // 204 - Xóa thành công
+            return Ok($"Xóa thành công bộ sưu tập {collection.Name}");
         }
 
         // 4. Hiển thị danh sách bộ sưu tập theo người dùng
-        [HttpGet("my-collection")]
+        [HttpGet("my-collections")]
         public async Task<IActionResult> GetCollectionsByUser()
         {
             //Kiểm tra token
